requirements.txt
# requirements.txt
streamlit>=1.32.0
tensorflow>=2.15.0
pillow>=10.2.0
numpy>=1.26.4
groq>=0.3.0
python-dotenv>=1.0.1